+++ test-app/src/domain/test/yaml/calculator.test.ts
@@ -1,0 +1,6 @@
+import { given } from "@epure/vitest/yaml";
+import { expect } from "vitest";
+
+given("a calculator", ({ left, right, result }) => {
+  expect(Number(left) + Number(right)).toBe(result);
+});